erro not found

+++ app/Http/Controllers/Admin/PlansController.php
@@ -53,10 +53,13 @@ class PlansController extends Controller
 
     public function destroy($url)
     {
-                $plan = $this->repository->where('url', $url)->first();
+                $plan = $this->repository->with('details')->where('url', $url)->first();
 
                 if(!$plan)
                 return redirect()->back();
+                if($plan->details->count() > 0) {
+                    return redirect()->back()->with('error', 'O plano esta vinculado a detalhes delete os detalhes antes');
+                }
 
                 $plan->delete();
                 return redirect()->route('plans.index',['plan' => $plan]);
